move outside of metadata folder
