+++ library/src/methods/getDescription/index.ts
