+++ library/src/methods/getMetadata/getMetadata.test-d.ts
@@ -1,7 +1,7 @@
 import { describe, expectTypeOf, test } from 'vitest';
-import { metadata } from '../../../actions/index.ts';
-import { string } from '../../../schemas/index.ts';
-import { pipe } from '../../pipe/index.ts';
+import { metadata } from '../../actions/index.ts';
+import { string } from '../../schemas/index.ts';
+import { pipe } from '../pipe/index.ts';
 import { getMetadata } from './getMetadata.ts';
 
 describe('getMetadata', () => {
