+++ library/src/methods/getMetadata/index.ts
