+++ library/src/methods/getTitle/getTitle.test-d.ts
@@ -1,7 +1,7 @@
 import { describe, expectTypeOf, test } from 'vitest';
-import { title } from '../../../actions/index.ts';
-import { string } from '../../../schemas/index.ts';
-import { pipe } from '../../pipe/index.ts';
+import { title } from '../../actions/index.ts';
+import { string } from '../../schemas/index.ts';
+import { pipe } from '../pipe/index.ts';
 import { getTitle } from './getTitle.ts';
 
 describe('getTitle', () => {
