+++ library/src/methods/getTitle/getTitle.ts
@@ -1,11 +1,11 @@
-import type { TitleAction } from '../../../actions/title/title.ts';
+import type { TitleAction } from '../../actions/title/title.ts';
 import {
   _createMetadataMethod,
   type UnknownPipe,
   type UnknownSchemaWithPipe,
-} from '../../../utils/_createMetadataMethod/_createMetadataMethod.ts';
-import type { SchemaWithPipe } from '../../pipe/pipe.ts';
-import type { SchemaWithPipeAsync } from '../../pipe/pipeAsync.ts';
+} from '../../utils/_createMetadataMethod/_createMetadataMethod.ts';
+import type { SchemaWithPipe } from '../pipe/pipe.ts';
+import type { SchemaWithPipeAsync } from '../pipe/pipeAsync.ts';
 
 export type ExtractTitle<
   TPipe extends UnknownPipe,
