+++ library/src/methods/index.ts
@@ -5,12 +5,14 @@ export * from './flatten/index.ts';
 export * from './forward/index.ts';
 export * from './getDefault/index.ts';
 export * from './getDefaults/index.ts';
+export * from './getDescription/index.ts';
 export * from './getFallback/index.ts';
 export * from './getFallbacks/index.ts';
+export * from './getMetadata/index.ts';
+export * from './getTitle/index.ts';
 export * from './is/index.ts';
 export * from './keyof/index.ts';
 export * from './message/index.ts';
-export * from './metadata/index.ts';
 export * from './omit/index.ts';
 export * from './parse/index.ts';
 export * from './parser/index.ts';
