--- library/src/methods/metadata/index.ts
@@ -1,3 +1,0 @@
-export * from './getDescription/index.ts';
-export * from './getMetadata/index.ts';
-export * from './getTitle/index.ts';
