+++ website/src/routes/api/(methods)/getDescription/properties.ts
