+++ website/src/routes/api/(methods)/getMetadata/properties.ts
